Refactor Layers screen; removing dependencies on Bayesian Network code. Change Drag Proxies and Factory Shapes to be centered around (0, 0) to support scaling for use in Palette and Layers panels"

+++ wires-core/wires-core-api/src/main/java/org/kie/wires/core/api/factories/ShapeDragProxy.java
@@ -33,8 +33,8 @@ public interface ShapeDragProxy<T extends Shape<T>> {
      * @param x
      * @param y
      */
-    void onDragEnd( final int x,
-                    final int y );
+    void onDragEnd( final double x,
+                    final double y );
 
     /**
      * Height of proxy used to ensure proxy is centered around mouse pointer position
